Add files via upload
we will learn to create a header section of a website, where we will add a navigation bar with a Logo, menu link, and a button. On this website background we will add a full-screen image with a dark color layer on the image. There will be a title, description, and search form on this website.